# lib/api-zod/src/generated/types/baselineRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceFingerprint } from "./deviceFingerprint";

export interface BaselineRecord {
  id: number;
  ip: string;
  label?: string | null;
  fingerprint: DeviceFingerprint;
  created_at: Date;
}