# lib/api-zod/src/generated/types/compareResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceComparison } from "./deviceComparison";

export interface CompareResult {
  total_packets: number;
  comparisons: DeviceComparison[];
}