# lib/api-zod/src/generated/types/detectionVerdict.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type DetectionVerdict =
  (typeof DetectionVerdict)[keyof typeof DetectionVerdict];

export const DetectionVerdict = {
  Normal: "Normal",
  Suspicious: "Suspicious",
  Possible_Spoofing: "Possible Spoofing",
  No_Baseline: "No Baseline",
} as const;