# lib/api-zod/src/generated/types/deviceComparison.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DetectionVerdict } from "./detectionVerdict";
import type { DeviceFingerprint } from "./deviceFingerprint";

export interface DeviceComparison {
  ip: string;
  verdict: DetectionVerdict;
  /** Overall similarity percentage (0-100) */
  similarity_score: number | null;
  baseline_label: string | null;
  current_fingerprint: DeviceFingerprint;
  baseline_fingerprint: DeviceFingerprint | null;
  /** Human-readable list of detected deviations */
  deviations: string[];
}