# lib/api-zod/src/generated/types/deviceFingerprint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceFingerprintProtocolDistribution } from "./deviceFingerprintProtocolDistribution";

export interface DeviceFingerprint {
  ip: string;
  packet_count: number;
  avg_packet_size: number;
  std_packet_size: number;
  avg_ttl: number;
  std_ttl: number;
  avg_time_delta: number;
  std_time_delta: number;
  protocol_distribution: DeviceFingerprintProtocolDistribution;
  unique_destinations: number;
}