# lib/api-zod/src/generated/types/saveBaselineRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceFingerprint } from "./deviceFingerprint";

export interface SaveBaselineRequest {
  ip: string;
  label?: string;
  fingerprint: DeviceFingerprint;
}